Moved Config.php out to AbstractConfig to make code DRY and help extn development
