+++ app/code/Magento/Ups/Helper/Config.php
@@ -5,41 +5,21 @@
  */
 namespace Magento\Ups\Helper;
 
+use Magento\Shipping\Model\ConfigInterface;
+
+
 /**
  * Configuration data of carrier
  */
-class Config
+class Config extends \Magento\Shipping\Helper\AbstractConfig implements ConfigInterface
 {
-    /**
-     * Get configuration data of carrier
-     *
-     * @param string $type
-     * @param string $code
-     * @return array|string|false
-     */
-    public function getCode($type, $code = '')
-    {
-        $codes = $this->getCodes();
-        if (!isset($codes[$type])) {
-            return false;
-        } elseif ('' === $code) {
-            return $codes[$type];
-        }
-
-        if (!isset($codes[$type][$code])) {
-            return false;
-        } else {
-            return $codes[$type][$code];
-        }
-    }
-
     /**
      * Get configuration data of carrier
      *
      * @return array
      * @SuppressWarnings(PHPMD.ExcessiveMethodLength)
      */
-    protected function getCodes()
+    function getCodes()
     {
         return [
             'action' => ['single' => '3', 'all' => '4'],
